support FixedSizeList Type Coercion
