+++ datafusion/sqllogictest/test_files/array.slt
@@ -4217,7 +4217,7 @@ NULL 10
 ## array_dims (aliases: `list_dims`)
 
 # array dims error
-query error Execution error: array_dims does not support type 'Int64'
+query error DataFusion error: Error during planning: No function matches the given name and argument types 'array_dims\(Int64\)'. You might need to add explicit type casts.\n\tCandidate functions:\n\tarray_dims\(array\)
 select array_dims(1);
 
 # array_dims scalar function
@@ -4328,6 +4328,7 @@ from array_ndims_table;
 ----
 1 1 2 5
 1 1 2 5
+NULL 1 2 5
 1 1 2 5
 
 query IIII
@@ -4340,6 +4341,7 @@ from large_array_ndims_table;
 ----
 1 1 2 5
 1 1 2 5
+NULL 1 2 5
 1 1 2 5
 
 statement ok
